# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Meta data analyzer API'
  version: '0.1'
paths:
  /api/webpages/analyze:
    post:
      tags:
        - 'Analyze Meta Data'
      summary: 'Analyze meta data of a webpage'
      operationId: analyze-metadata
      requestBody:
        description: 'Analyze webpage meta data'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebpageDto'
      responses:
        400:
          description: 'Bad Request'
  /api/webpages/preview:
    post:
      tags:
        - 'Create Preview Image'
      summary: 'Create preview image of a webpage'
      operationId: preview-image
      requestBody:
        description: 'Create preview image'
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebpageDto'
      responses:
        400:
          description: 'Bad Request'
  /api/webpages/paginate:
    get:
      tags:
        - 'Get all webpage and metadata'
      summary: 'Get all webpage and metadata info with pagination'
      operationId: webpage-metadata
      parameters:
        -
          name: page
          in: query
          description: 'Page number'
          required: true
      responses:
        400:
          description: 'Bad Request'
  /api/webpages/metadata/update-all:
    get:
      tags:
        - 'Update all metadata'
      summary: 'Update all metadata'
      operationId: update-all-metadata
      responses:
        400:
          description: 'Bad Request'
  '/api/webpages/preview/{id}':
    get:
      tags:
        - 'Display preview image'
      summary: 'Display preview image by webpage id'
      operationId: display-preview-image
      parameters:
        -
          name: id
          in: path
          description: 'Webpage id'
          required: true
      responses:
        400:
          description: 'Bad Request'
components:
  schemas:
    WebpageDto:
      title: 'Web Page Request dto'
      description: 'Web Page Request dto'
      properties:
        url:
          title: URL
          description: 'A URL need to capture all meta data and analyze or create preview image.'
          type: string
      type: object